refactor: token usage styles
